Fix change password.
