+++ src/pages/change-password/index.js
@@ -5,15 +5,8 @@ import { AuthContext } from 'src/providers/AuthContext'
 import { useContext } from 'react'
 
 const ChangePass = () => {
-  const { encryptePWD, loggedUser } = useContext(AuthContext)
-
   return (
-    <ImmediateWindow
-      datasetId={ResourceIds.ChangePassword}
-      props={(encryptePWD, loggedUser)}
-      titleName={'ChangePassword'}
-      Component={ChangePassword}
-    />
+    <ImmediateWindow datasetId={ResourceIds.ChangePassword} titleName={'ChangePassword'} Component={ChangePassword} />
   )
 }
